Add files via upload
Here I countiued to work on the game part of the program, where the interactive algorithm builds up suspense to get the answer. So far, it found if your answer is right or not but I will develop it to make use of more doubtful methods, to really test the user to see if he or she knows what she or he is supposed to know. There is now software adavatage in doing this it is only for the user to know.